refactor: move ui/ to internal/ui/
